#3 Fixed and cleaned up the in-game scene and scripts
