--- Assets/teleportation.cs
@@ -1,7 +1,0 @@
-namespace DefaultNamespace
-{
-    public class teleportation
-    {
-        
-    }
-}
