--- Assets/teleportation.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 842a87157b7e4f549793ee41c418b7c5
-timeCreated: 1556988741
